This patch explicitly defines how to extract such comments from JavaScript, CSS and WebAssembly sources.
It defines multiple ways to do so: either by actually parsing the code,
or by just going through all the lines of the program looking for what
"looks like" a comment. This is so that different implementations can
choose what's best for them, depending on whether they are already
parsing the code or not.

To ensure consist behavior accross implementations that choose different
strategies, the specification enforces additional requirements on
tools that append a `sourceMappingURL` comment to the generated code:
the comment must be placed in such a way that all extraction methods
yield the same result. This is not an unresonable burden, since
if the progeram is syntactically valid, simply adding the comment at the
end of the file only potentially followed by other tool-injected
comments is enough. This requirement is lifted if the input code given
to the tool is already "maliciously crafted", since we would otherwise
require tool to go rewrite that code (for example, splitting strings
that contain something that looks like a comment).

I have left the CSS extraction method as TODO because first I want to
check how do you feel about the JS one. It has the following properties:
- It iterates line by line. Implementations can thus optimize it by
  going through each line _in reverse order_, and then scanning through
  its characters from the beginning to the end (which is what a regexp
  would do).
- It expects multi-line comments to actually be in a single line.
- It returns the last `sourceMappingURL` comment (or well, comment-like)
  found in the source.
- It only considers comments after the last piece of code (i.e. it
  discards any comment found so far every time it sees some non-comment
  non-whitespace characters).
- It has no requirements about what is _before_ a comment. Adding the
  comment at the end of the file without first ensuring that there is
  a newline before it is valid.