Update `binding` expression parsing
This changes the order of binding expressions to use encode `M` first. If `M` is >= -1, then `M` is treated as an offset into `names` for the entirety of the generated range. If it's negative, then there are `-M` sub-ranges for this variable within the generated range, and the next VLQ is the offset into `names` to use for the start of the generated range.