Add additional constraint to ensure mappings stay between offsets
An implementation may do a generated->original lookup in an index map
by first searching through the sections by their offset and looking in
only the section where the position is within the range {offset, nextOffset}.
This is only valid if all mappings in a given section are contained in
that range, rather than going into the next section's range.

Also simplifies the check by removing the firstPreviousMapping variable
and its use.

The algorithm checks that the
  (1) the offsets are ordered (in 8g)
  (2) the last mapping of the previous section is before the
      current offset (in 8h)
  (3) it doesn't check that the first mapping is after the
      current offset because that is true by construction